fix by comments
